+++ example/src/main/java/com/umang/example/fcmclient/ExampleApplication.kt
@@ -1,6 +1,7 @@
 package com.umang.example.fcmclient
 
 import android.app.Application
+import android.os.StrictMode
 import com.umang.fcmclient.FcmClientHelper
 import com.umang.fcmclient.util.Logger
 
@@ -9,10 +10,17 @@ import com.umang.fcmclient.util.Logger
  */
 class ExampleApplication : Application() {
     override fun onCreate() {
+        StrictMode.setThreadPolicy(
+            StrictMode.ThreadPolicy.Builder()
+                .detectDiskReads()
+                .detectDiskWrites()
+                .detectNetwork() // or .detectAll() for all detectable problems
+                .penaltyLog()
+                .build()
+        )
         super.onCreate()
         // initialize sdk
-        FcmClientHelper.getInstance(applicationContext).initialise(this, Logger.LogLevel.VERBOSE,
-                4)
+        FcmClientHelper.getInstance(applicationContext).initialise( Logger.LogLevel.VERBOSE, 4)
         FcmClientHelper.getInstance(applicationContext).addListener(FcmListener1())
         FcmClientHelper.getInstance(applicationContext).addListener(FcmListener2())
     }