+++ fcm-client/src/main/java/com/umang/fcmclient/FcmClientHelper.kt
@@ -2,11 +2,11 @@ package com.umang.fcmclient
 
 import android.app.Application
 import android.content.Context
-import com.google.firebase.iid.FirebaseInstanceId
+import androidx.lifecycle.ProcessLifecycleOwner
 import com.google.firebase.messaging.FirebaseMessaging
 import com.google.firebase.messaging.RemoteMessage
 import com.umang.fcmclient.listeners.FirebaseMessageListener
-import com.umang.fcmclient.repository.Repository
+import com.umang.fcmclient.repository.Provider
 import com.umang.fcmclient.util.Logger
 import com.umang.fcmclient.util.isDebugBuild
 import java.util.concurrent.Executors
@@ -20,16 +20,11 @@ public class FcmClientHelper internal constructor(private var context: Context) 
 
     private val logger = Logger.getLogger("FcmClientHelper")
 
-    private var activityCounter = 0
-
     private val listeners = mutableListOf<FirebaseMessageListener>()
 
     private var scheduledExecutor = Executors.newScheduledThreadPool(1)
 
-    private val isAppInForeground: Boolean
-        get() = activityCounter > 0
-
-    private val repository = Repository(context)
+    private var isAppInForeground: Boolean = false
 
     private var retryInterval: Long = 30
 
@@ -63,19 +58,46 @@ public class FcmClientHelper internal constructor(private var context: Context) 
      * which the SDK should retry registering for Push Token in case of any failure. Default
      * value is 30 seconds.
      *
-     * ```
-     *
-     * FCMClientHelper.getInstance(applicationContext).initialise(this)
-     *
-     * ```
      */
-    public fun initialise(application: Application, logLevel: Logger.LogLevel = Logger.LogLevel.ERROR,
-                   retryInterval: Long = 30) {
+    @Deprecated(
+        "This API has been deprecated since 2.1.00",
+        ReplaceWith(
+            "FcmClientHelper.getInstance(applicationContext).initialise()",
+            "com.umang.fcmclient.FcmClientHelper"
+        ),
+        DeprecationLevel.WARNING
+    )
+    public fun initialise(
+        application: Application,
+        logLevel: Logger.LogLevel = Logger.LogLevel.ERROR,
+        retryInterval: Long = 30
+    ) {
         try {
-            application.registerActivityLifecycleCallbacks(ActivityLifecycleCallbacks())
+            initialise(logLevel, retryInterval)
+        } catch (e: Exception) {
+            logger.error(" initialise() ", e)
+        }
+    }
+
+    /**
+     * Initialize the FCM client. This needs to be called from the onCreate() of [Application] class.
+     *
+     * @param logLevel optional parameter takes in the level of logs which should be printed by the
+     * SDK when the application is running in debug mode. By default, only error logs are printed.
+     * Refer to [Logger.LogLevel] for more details.
+     * @param retryInterval optional parameter which takes the time interval(in seconds) after
+     * which the SDK should retry registering for Push Token in case of any failure. Default
+     * value is 30 seconds.
+     */
+    public fun initialise(
+        logLevel: Logger.LogLevel = Logger.LogLevel.ERROR,
+        retryInterval: Long = 30
+    ) {
+        try {
+            ProcessLifecycleOwner.get().lifecycle.addObserver(AppLifecycleObserver(context.applicationContext))
             Logger.logLevel = logLevel
-            Logger.isLogEnabled = isDebugBuild(application.applicationContext)
-            if (retryInterval >= 5){
+            Logger.isLogEnabled = isDebugBuild(context.applicationContext)
+            if (retryInterval >= 5) {
                 this.retryInterval = retryInterval
             }
             logger.verbose(" initialise() Initialising fcm client library. Log level - $logLevel")
@@ -105,6 +127,9 @@ public class FcmClientHelper internal constructor(private var context: Context) 
                 for (topic in topics) {
                     logger.verbose("Subscribing to $topic")
                     FirebaseMessaging.getInstance().subscribeToTopic(topic)
+                        .addOnCompleteListener { task ->
+                            logger.verbose("subscribeToTopic() isSuccess ${task.isSuccessful}")
+                        }
                 }
             } catch (e: Exception) {
                 logger.error(" subscribeToTopic() ", e)
@@ -129,61 +154,55 @@ public class FcmClientHelper internal constructor(private var context: Context) 
         }
     }
 
-    internal fun onStart() {
-        try {
-            if (activityCounter == 0) {
-                registerForPush()
-            }
-            activityCounter++
-            logger.verbose("onStart(): activity counter: $activityCounter")
-        } catch (e: Exception) {
-            logger.error(" onStart() ", e)
-        }
-    }
-
-    internal fun onStop() {
-        try {
-            activityCounter--
-            logger.verbose("onStop(): activity counter: $activityCounter")
-            if (activityCounter == 0) {
-                onAppBackground()
-            }
-        } catch (e: Exception) {
-            logger.error(" onStop() ", e)
-        }
-    }
-
     internal fun onNewToken(token: String) {
-        try {
-            notifyListenersIfRequired(token)
-            repository.saveToken(token)
-        } catch (e: Exception) {
-            logger.error(" onNewToken() ", e)
+        AsyncExecutor.submit {
+            try {
+                Provider.getRepository(context).saveToken(token)
+                notifyListeners(token)
+            } catch (e: Exception) {
+                logger.error(" onNewToken() ", e)
+            }
         }
     }
 
     private fun registerForPushIfRequired() {
-        val savedToken: String = repository.getToken()
+        val savedToken = Provider.getRepository(context).getToken()
         if (savedToken.isEmpty()) {
             registerForPush()
         }
     }
 
-    private fun onAppBackground() {
-        shutdownRetryScheduler()
-        registerForPushIfRequired()
+    internal fun onAppForeGround() {
+        try {
+            logger.verbose("onAppForeGround(): Application coming to foreground.")
+            registerForPush()
+            isAppInForeground = true
+        } catch (e: Exception) {
+            logger.error(" onAppForeGround(): Exception: ", e)
+        }
+    }
+
+    internal fun onAppBackground() {
+        try {
+            logger.verbose("onAppBackground(): Application going to background.")
+            isAppInForeground = false
+            shutdownRetryScheduler()
+            registerForPushIfRequired()
+        } catch (e: Exception) {
+            logger.error(" onAppBackground(): Exception: ", e)
+        }
     }
 
     private fun registerForPush() {
         logger.verbose(" registerForPush(): Will register for push.")
-        FirebaseInstanceId.getInstance().instanceId.addOnCompleteListener { task ->
+        FirebaseMessaging.getInstance().token.addOnCompleteListener { task ->
             try {
                 if (!task.isSuccessful) {
                     logger.verbose(" registerForPush(): Token registration failed.")
                     scheduleRetry()
                     return@addOnCompleteListener
                 }
-                val token = task.result?.token
+                val token = task.result
                 if (token.isNullOrEmpty()) {
                     logger.verbose(" registerForPush(): Token null or empty.")
                     scheduleRetry()
@@ -208,28 +227,24 @@ public class FcmClientHelper internal constructor(private var context: Context) 
     }
 
     private fun shutdownRetryScheduler() {
-        if (!scheduledExecutor.isShutdown){
+        if (!scheduledExecutor.isShutdown) {
             logger.verbose(" shutdownRetryScheduler() Shutting down retry scheduler.")
             scheduledExecutor.shutdownNow()
         }
     }
 
-    private fun notifyListenersIfRequired(token: String) {
-        AsyncExecutor.submit {
-            try {
-                val savedToken = repository.getToken()
-                if (token == savedToken) return@submit
-                logger.verbose(" notifyListenersIfRequired() : Notifying listeners")
-                for (listener in listeners) {
-                    try {
-                        listener.onTokenAvailable(token)
-                    } catch (e: Exception) {
-                        logger.error(" notifyListenersIfRequired() : ", e)
-                    }
+    private fun notifyListeners(token: String) {
+        try {
+            logger.verbose(" notifyListenersIfRequired() : Notifying listeners")
+            for (listener in listeners) {
+                try {
+                    listener.onTokenAvailable(token)
+                } catch (e: Exception) {
+                    logger.error(" notifyListenersIfRequired() : ", e)
                 }
-            } catch (e: Exception) {
-                logger.error(" notifyListenersIfRequired() ", e)
             }
+        } catch (e: Exception) {
+            logger.error(" notifyListenersIfRequired() ", e)
         }
     }
 
@@ -242,12 +257,11 @@ public class FcmClientHelper internal constructor(private var context: Context) 
          * @return instance of [FcmClientHelper]
          */
         public fun getInstance(context: Context): FcmClientHelper {
-            if (instance == null) {
-                synchronized(FcmClientHelper::class.java) {
-                    if (instance == null) instance = FcmClientHelper(context)
-                }
+            return instance ?: synchronized(FcmClientHelper::class.java){
+                val inst = instance ?: FcmClientHelper(context)
+                instance = inst
+                inst
             }
-            return instance as FcmClientHelper
         }
     }
 }